+++ time-hourglass.cabal
@@ -63,6 +63,7 @@ library
   default-language: Haskell2010
   if os(windows)
     other-modules:
+        Time.Calendar.FromUnixEpoch
         Time.System.OS
     hs-source-dirs:
         src/win
@@ -70,7 +71,9 @@ library
         Win32
   else
     other-modules:
+        Time.Calendar.FromUnixEpoch
         Time.System.OS
+        Time.System.Types
     hs-source-dirs:
         src/unix
     other-extensions:
